# src/docs/swagger.yaml
openapi: "3.0.2"
info:
  version: 1.0.0
  title: Notes Taking Application
servers:
  - url: http://localhost:3000/api
...
paths:
...
  /note/notes:
    get:
      summary: find all notes
      operationId: findAllNotes
      tags:
        - Notes
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/headerPlatform"
        - $ref: "#/components/parameters/headerVersion"
      responses:
        200:
          description: all notes
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Notes"
        400:
          $ref: "#/components/responses/BadRequestError"
    post:
      summary: create a note
      operationId: createNotes
      tags:
        - Notes
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/headerPlatform"
        - $ref: "#/components/parameters/headerVersion"
      requestBody:
        required: true
        content:
          application/json:
            example:
              title: "Test Title"
              content: "Notes in details"
              tags: ""
            schema:
              type: object
              required:
                - title
                - content
                - tags
              properties:
                title:
                  type: string
                content:
                  type: string
                tags:
                  type: string
      responses:
        200:
          description: note created!!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        400:
          $ref: "#/components/responses/BadRequestError"
...
components:
  parameters:
    headerPlatform:
      name: X-Notes-Platform
      in: header
      required: true
      schema:
        type: string
        default: ios
        enum:
          - ios
          - android
    headerVersion:
      name: X-Notes-Version
      in: header
      required: true
      schema:
        type: string
        default: 1.0.0
        pattern: '^[\d]+\.[\d]+\.[\d]+$'
...
  schemas:
    ApiResponse:
      required:
        - success
        - message
      properties:
        success:
          type: boolean
        message:
          type: string
        data: {}
...
    Notes:
      properties:
        title:
          type: string
        content:
          type: string
          format: email
        tags:
          type: string
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
  responses:
    BadRequestError:
      description: Request fails validation or doesn't meet all conditions for request
      content:
        application/json:
          schema:
            example:
              success: false
              message: "Validation failed"
              data: ""
            allOf:
              - $ref: "#/components/schemas/ApiResponse"
              - type: object
                properties:
                  data:
                    type: object
                    nullable: true
...